+++ acceptance/src/App.js
@@ -4,7 +4,6 @@ import { BrowserRouter as Router, Route } from 'react-router-dom';
 
 import routes from './routes';
 import Nav from './components/Nav';
-import "@hig/styles/build/index.css";
 
 class App extends Component {
   render() {