+++ acceptance/src/routes.js
@@ -4,8 +4,10 @@ import ButtonsPage from "./pages/ButtonsPage";
 import FlyoutPage from "./pages/FlyoutPage";
 import CheckboxPage from "./pages/CheckboxPage";
 import DropdownPage from "./pages/DropdownPage";
+import IconnButtonsPage from "./pages/IconButtonsPage";
 import IconsPage from "./pages/IconsPage";
 import InputPage from "./pages/InputPage";
+import LabelPage from "./pages/LabelPage";
 import ModalPage from "./pages/ModalPage";
 import ProgressBarPage from "./pages/ProgressBarPage";
 import ProgressRingPage from "./pages/ProgressRingPage";
@@ -26,7 +28,9 @@ export default [
   { path: "/flyout/", component: FlyoutPage, label: "Flyout" },
   { path: "/checkbox/", component: CheckboxPage, label: "Checkbox" },
   { path: "/icons/", component: IconsPage, label: "Icons" },
+  { path: "/icon-button/", component: IconnButtonsPage, label: "Icon Button" },
   { path: "/input/", component: InputPage, label: "Input" },
+  { path: "/label/", component: LabelPage, label: "Label"},
   { path: "/modal/", component: ModalPage, label: "Modal" },
   { path: "/progress-bar/", component: ProgressBarPage, label: "Progress Bar" },
   { path: "/progress-ring/", component: ProgressRingPage, label: "Progress Ring" },