+++ packages/project-account-switcher/src/__gemini__/ProjectAccountSwitcher.stories-test.js
@@ -1,5 +1,7 @@
 import React from "react";
 import { storiesOf } from "@storybook/react";
+import accountImage from "@hig/storybook/storybook-support/fixtures/account/dam.png";
+import projectImage from "@hig/storybook/storybook-support/fixtures/project/architecture.png";
 
 import ProjectAccountSwitcher from "../ProjectAccountSwitcher";
 
@@ -36,18 +38,27 @@ storiesOf("ProjectAccountSwitcher", module)
   .add("open-accounts-projects", () => (
     <Capture>
       <ProjectAccountSwitcher
-        accounts={[{ id: "1", label: "Account 1" }]}
+        accounts={[
+          { id: "1", label: "Account 1" },
+          { id: "2", label: "Account 2", image: accountImage }
+        ]}
         activeAccountId="1"
         activeProjectId="1"
         open
-        projects={[{ id: "1", label: "Project 1" }]}
+        projects={[
+          { id: "1", label: "Project 1" },
+          { id: "2", label: "Project 2", image: projectImage }
+        ]}
       />
     </Capture>
   ))
   .add("open-accounts", () => (
     <Capture>
       <ProjectAccountSwitcher
-        accounts={[{ id: "1", label: "Account 1" }]}
+        accounts={[
+          { id: "1", label: "Account 1" },
+          { id: "2", label: "Account 2", image: accountImage }
+        ]}
         activeAccountId="1"
         open
       />
@@ -58,7 +69,10 @@ storiesOf("ProjectAccountSwitcher", module)
       <ProjectAccountSwitcher
         activeProjectId="1"
         open
-        projects={[{ id: "1", label: "Project 1" }]}
+        projects={[
+          { id: "1", label: "Project 1" },
+          { id: "2", label: "Project 2", image: projectImage }
+        ]}
       />
     </Capture>
   ));