+++ packages/side-nav/src/Link/index.js
@@ -1,3 +1,1 @@
-import "@hig/styles/build/fonts.css";
-
 export { default } from "./Link";