+++ packages/theme-data/src/colorSchemes/darkBlue/components/checkbox.js
@@ -14,14 +14,6 @@ export default {
   },
 
   // Focus
-  "checkbox.focus.borderColor": {
-    value: {
-      ref: "basics.colors.white"
-    },
-    transform: {
-      alpha: 0.5
-    }
-  },
   "checkbox.focus.halo.color": {
     value: {
       ref: "basics.colors.autodeskBlue400"
@@ -46,16 +38,6 @@ export default {
     },
     transform: {
       alpha: 0.15
-    }
-  },
-
-  // Pressed
-  "checkbox.pressed.halo.color": {
-    value: {
-      ref: "basics.colors.autodeskBlue400"
-    },
-    transform: {
-      alpha: 0.35
     }
   },