+++ packages/theme-data/src/colorSchemes/darkGray/components/button.js
@@ -1,0 +1,12 @@
+export default {
+  "button.outline.pressed.halo.color": {
+    value: {
+      ref: "button.solid.pressed.halo.color"
+    }
+  },
+  "button.flat.pressed.halo.color": {
+    value: {
+      ref: "button.solid.pressed.halo.color"
+    }
+  }
+};