+++ packages/components/CHANGELOG.md
@@ -1,3 +1,22 @@
+# [@hig/components-v2.0.1](https://github.com/Autodesk/hig/compare/@hig/components@2.0.0...@hig/components@2.0.1) (2019-07-16)
+
+
+### Bug Fixes
+
+* update icons package ([406803c](https://github.com/Autodesk/hig/commit/406803c))
+
+# [@hig/components-v2.0.0](https://github.com/Autodesk/hig/compare/@hig/components@1.1.2...@hig/components@2.0.0) (2019-05-10)
+
+
+### Features
+
+* remove table component ([18bae46](https://github.com/Autodesk/hig/commit/18bae46))
+
+
+### BREAKING CHANGES
+
+* * the table component is removed because it is no longer supported
+
 # [@hig/components-v1.1.2](https://github.com/Autodesk/hig/compare/@hig/components@1.1.1...@hig/components@1.1.2) (2019-04-24)