+++ packages/side-nav/CHANGELOG.md
@@ -1,3 +1,25 @@
+# [@hig/side-nav-v2.1.1](https://github.com/Autodesk/hig/compare/@hig/side-nav@2.1.0...@hig/side-nav@2.1.1) (2019-07-16)
+
+
+### Bug Fixes
+
+* external link icon colors ([c980876](https://github.com/Autodesk/hig/commit/c980876))
+* update SideNav to newest Icons version ([d16d999](https://github.com/Autodesk/hig/commit/d16d999))
+
+# [@hig/side-nav-v2.1.0](https://github.com/Autodesk/hig/compare/@hig/side-nav@2.0.5...@hig/side-nav@2.1.0) (2019-06-12)
+
+
+### Features
+
+* Enable onClick events for sideNav headerLink and superHeaderLink ([a2c1d22](https://github.com/Autodesk/hig/commit/a2c1d22))
+
+# [@hig/side-nav-v2.0.5](https://github.com/Autodesk/hig/compare/@hig/side-nav@2.0.4...@hig/side-nav@2.0.5) (2019-05-10)
+
+
+### Bug Fixes
+
+* remove unused styles dependency ([60d20f3](https://github.com/Autodesk/hig/commit/60d20f3))
+
 # [@hig/side-nav-v2.0.4](https://github.com/Autodesk/hig/compare/@hig/side-nav@2.0.3...@hig/side-nav@2.0.4) (2019-04-24)